requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==3.0.1
pymysql==1.1.0
cryptography==41.0.7
stripe==7.8.0
python-dotenv==1.0.0
email-validator==2.1.0